+++ Assets/Code/Scripts/Player.cs
@@ -2,9 +2,9 @@
 {
     public class Player
     {
-	    private int _atk; // 공격력
-		private int _hp; //체력
+	    private float _atk; // 공격력
+		private float _hp; //체력
 		private int _lv; // 레벨
-		private int _exp; // 경험치
+		private float _exp; // 경험치
     }
 }
